Fixed the code generator and added a new feature
The code generation was not returning the numeric columns correctly
which was fixed, Fixed issues related to separation of the EmailInfo
from the Site Entity. Fixed an issue with entity processing and the
database resultset cursor, calling first() ensured the cursor pointed to
the first record rather than one before the first record.